LayoutTests/js/dom/dom-properties-are-configurable.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test-pre.js"></script>
</head>
<script>
    description("Verify that certain DOM properties are configurable");

    // Some properties known to cause issues (see rdar://problem/20807563).
    shouldBeTrue('Object.getOwnPropertyDescriptor(HTMLElement.prototype, "innerText").configurable');
    shouldBeTrue('Object.getOwnPropertyDescriptor(window, "onhashchange").configurable');
    shouldBe('HTMLElement.prototype.__defineGetter__("innerText", function() { return "NO!" });', 'undefined');
    shouldBe('window.__defineGetter__("onhashchange", function() { return "WebKit!" });', 'undefined');
    shouldBeEqualToString('HTMLElement.prototype.innerText', 'NO!');
    shouldBeEqualToString('document.documentElement.innerText', 'NO!');
    shouldBeEqualToString('window.onhashchange', 'WebKit!');
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
